requirements.txt
altair==4.2.2
appnope==0.1.3
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
bcrypt==4.0.1
blinker==1.6.2
branca==0.6.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
extra-streamlit-components==0.1.56
folium==0.14.0
fonttools==4.39.3
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
importlib-metadata==6.4.0
importlib-resources==5.12.0
ipykernel==6.22.0
ipython==8.12.0
japanize-matplotlib==1.1.3
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
nest-asyncio==1.5.6
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.1
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pydeck==0.8.1b0
Pygments==2.15.0
PyJWT==2.6.0
Pympler==1.0.1
pyparsing==3.0.9
pyproj==3.5.0
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==25.0.2
requests==2.28.2
rich==13.3.4
scipy==1.10.1
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
streamlit==1.21.0
streamlit-authenticator==0.2.1
streamlit-folium==0.11.1
toml==0.10.2
toolz==0.12.0
tornado==6.2
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
validators==0.20.0
wcwidth==0.2.6
zipp==3.15.0
